Broken Link seperated into model and service
